templates/backend/python/fastapi/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
aiosmtplib==3.0.1
jinja2==3.1.3

# Database - SQLAlchemy (for SQL databases)
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pymysql==1.1.0
aiosqlite==0.19.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# OAuth (optional - installed conditionally)
# authlib==1.3.0
# httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
